Cleanup chain objects and configs into the chain api crate
