+++ chain/api/Cargo.toml
@@ -18,10 +18,15 @@ async-std = { workspace = true }
 async-trait = { workspace = true }
 async-lock = { workspace = true }
 futures = { workspace = true }
+semver = "1.0"
 serde = { workspace = true }
+serde_json = { workspace = true }
+serde_with = { workspace = true }
 smart-default = { workspace = true }
+strum = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
+validator = { workspace = true }
 
 hopr-crypto-types = { workspace = true }
 chain-actions = { workspace = true }
