+++ chain/api/src/errors.rs
@@ -15,6 +15,9 @@ pub enum HoprChainError {
 
     #[error(transparent)]
     DbError(#[from] hopr_db_api::errors::DbError),
+
+    #[error("configuration error: {0}")]
+    Configuration(String),
 }
 
 /// The default [Result] object translating errors in the [HoprChainError] type
