+++ chain/api/src/lib.rs
@@ -1,9 +1,16 @@
 //! Crate containing the API object for chain operations used by the HOPRd node.
 
+pub mod config;
 pub mod errors;
 pub mod executors;
 
+use chain_actions::action_queue::{ActionQueue, ActionQueueConfig};
+use chain_actions::action_state::IndexerActionTracker;
+use chain_actions::payload::SafePayloadGenerator;
+use chain_rpc::rpc::RpcOperationsConfig;
 pub use chain_types::chain_events::SignificantChainEvent;
+use config::ChainNetworkConfig;
+use executors::{EthereumTransactionExecutor, RpcEthereumClient, RpcEthereumClientConfig};
 pub use hopr_internal_types::channels::ChannelEntry;
 
 use async_lock::RwLock;
@@ -12,8 +19,8 @@ use std::time::Duration;
 
 use chain_actions::ChainActions;
 use chain_indexer::{block::Indexer, handlers::ContractEventHandlers, IndexerConfig};
-use chain_rpc::rpc::RpcOperations;
 use chain_rpc::HoprRpcOperations;
+use chain_rpc::{rpc::RpcOperations, TypedTransaction};
 use chain_types::ContractAddresses;
 use hopr_crypto_types::prelude::*;
 use hopr_internal_types::account::AccountEntry;
@@ -100,6 +107,84 @@ pub async fn wait_for_funds<Rpc: HoprRpcOperations>(
     }
 
     Err(HoprChainError::Api("timeout waiting for funds".into()))
+}
+
+type ActiveTxExecutor = EthereumTransactionExecutor<
+    TypedTransaction,
+    RpcEthereumClient<RpcOperations<JsonRpcClient>>,
+    SafePayloadGenerator,
+>;
+
+pub fn build_chain_components<Db>(
+    me_onchain: &ChainKeypair,
+    chain_config: ChainNetworkConfig,
+    contract_addrs: ContractAddresses,
+    module_address: Address,
+    db: Db,
+) -> (
+    ActionQueue<Db, IndexerActionTracker, ActiveTxExecutor>,
+    ChainActions<Db>,
+    RpcOperations<JsonRpcClient>,
+)
+where
+    Db: HoprDbAllOperations + Clone + Send + Sync + std::fmt::Debug + 'static,
+{
+    // TODO: extract this from the global config type
+    let rpc_http_config = chain_rpc::client::native::HttpPostRequestorConfig::default();
+
+    // TODO: extract this from the global config type
+    let rpc_http_retry_policy = SimpleJsonRpcRetryPolicy {
+        min_retries: Some(2),
+        ..SimpleJsonRpcRetryPolicy::default()
+    };
+
+    // TODO: extract this from the global config type
+    let rpc_cfg = RpcOperationsConfig {
+        chain_id: chain_config.chain.chain_id as u64,
+        contract_addrs,
+        module_address,
+        expected_block_time: Duration::from_millis(chain_config.chain.block_time),
+        tx_polling_interval: Duration::from_millis(chain_config.tx_polling_interval),
+        finality: chain_config.confirmations,
+        max_block_range_fetch_size: chain_config.max_block_range,
+    };
+
+    // TODO: extract this from the global config type
+    let rpc_client_cfg = RpcEthereumClientConfig::default();
+
+    // TODO: extract this from the global config type
+    let action_queue_cfg = ActionQueueConfig::default();
+
+    // --- Configs done ---
+
+    // Build JSON RPC client
+    let rpc_client = JsonRpcClient::new(
+        &chain_config.chain.default_provider,
+        DefaultHttpPostRequestor::new(rpc_http_config),
+        rpc_http_retry_policy,
+    );
+
+    // Build RPC operations
+    let rpc_operations = RpcOperations::new(rpc_client, me_onchain, rpc_cfg).expect("failed to initialize RPC");
+
+    // Build the Ethereum Transaction Executor that uses RpcOperations as backend
+    let ethereum_tx_executor = EthereumTransactionExecutor::new(
+        RpcEthereumClient::new(rpc_operations.clone(), rpc_client_cfg),
+        SafePayloadGenerator::new(me_onchain, contract_addrs, module_address),
+    );
+
+    // Build the Action Queue
+    let action_queue = ActionQueue::new(
+        db.clone(),
+        IndexerActionTracker::default(),
+        ethereum_tx_executor,
+        action_queue_cfg,
+    );
+
+    // Instantiate Chain Actions
+    let chain_actions = ChainActions::new(me_onchain.public().to_address(), db, action_queue.new_sender());
+
+    (action_queue, chain_actions, rpc_operations)
 }
 
 /// Repsents all chain interactions exported to be used in the hopr-lib
