+++ hopr/hopr-lib/Cargo.toml
@@ -19,9 +19,7 @@ const_format = "0.2.32"
 futures = { workspace = true }
 lazy_static = { workspace = true }
 tracing = { workspace = true }
-semver = "1.0"
 serde = { workspace = true }
-serde_json = { workspace = true }
 serde_with = { workspace = true }
 strum = { workspace = true }
 smart-default = { workspace = true }
