+++ hopr/hopr-lib/src/config.rs
@@ -46,7 +46,7 @@ pub struct Chain {
     #[serde(default)]
     pub provider: Option<String>,
     #[serde(default)]
-    pub protocols: crate::chain::ProtocolsConfig,
+    pub protocols: chain_api::config::ProtocolsConfig,
     #[serde(default = "just_true")]
     #[default = true]
     pub check_unrealized_balance: bool,
