+++ hoprd/hoprd/src/bin/hoprd-cfg.rs
@@ -82,14 +82,17 @@ fn main() -> Result<(), hoprd::errors::HoprdError> {
             .hopr
             .chain
             .protocols
-            .supported_networks()
+            .supported_networks(hopr_lib::constants::APP_VERSION_COERCED)
             .iter()
             .any(|network| network == &cfg.hopr.chain.network)
         {
             return Err(hoprd::errors::HoprdError::ValidationError(format!(
                 "The specified network '{}' is not listed as supported ({:?})",
                 cfg.hopr.chain.network,
-                cfg.hopr.chain.protocols.supported_networks()
+                cfg.hopr
+                    .chain
+                    .protocols
+                    .supported_networks(hopr_lib::constants::APP_VERSION_COERCED)
             )));
         };
